+++ src/types/index.ts
@@ -1,10 +1,10 @@
-export { ICompile } from './compile';
-export { ICompileUtilForRepeat, ICompileUtil } from './compile-utils';
+export { ICompile } from './platform-browser/compile';
+export { ICompileUtilForRepeat, ICompileUtil } from './platform-browser/compile-utils';
 export { ComponentList, IComponent, SetState } from './component';
 export { IMiddleware, EsRouteObject, IInDiv } from './indiv';
 export { INvModule, TInjectTokenProvider, TUseClassProvider, TuseValueProvider } from './nv-module';
 export { IKeyWatcher } from './key-watcher';
-export { TRouter, IRouter, GetLocation, SetLocation } from './router';
+export { TRouter, IRouter, GetLocation, SetLocation } from './platform-browser/router';
 export { IUtil } from './utils';
-export { IVnode, TAttributes, IPatchList } from './virtual-dom';
+export { IVnode, TAttributes, IPatchList } from './platform-browser/virtual-dom';
 export { TFnWatcher, TFnRender, IWatcher } from './watcher';
